+++ src/sw/allotmentclub/browser/tests/test_account.py
@@ -107,7 +107,8 @@ def test_SEPASammlerExportView_1(browser):
         SEPASammlerEntry.find_or_create(
             sepasammler=sammler, value=value,
             member=Member.create(
-                lastname='Müller', direct_debit=True, iban=iban, bic='NOLADE',
+                lastname='Müller', direct_debit=True, iban=iban,
+                bic='NOLADE21HAL',
                 direct_debit_date=datetime.date(2017, 1, 1)))
     setUp()
     browser.login()
@@ -115,8 +116,8 @@ def test_SEPASammlerExportView_1(browser):
         'http://localhost/accounts/sepa_sammler/{}/export'.format(sammler.id))
     doc = lxml.etree.fromstring(browser.contents.encode('utf-8'))
     assert '1075.39' == doc.find('.//CtrlSum', namespaces=doc.nsmap).text
-    assert 'Müller, ' == doc.findall('.//Nm', namespaces=doc.nsmap)[-1].text
-    assert 'Mueller-' == doc.find('.//MndtId', namespaces=doc.nsmap).text
+    assert 'Mueller, ' == doc.findall('.//Nm', namespaces=doc.nsmap)[-1].text
+    assert 'MUELLER' == doc.find('.//MndtId', namespaces=doc.nsmap).text
 
 
 def test_BankingAccountListReportView_1(browser):